conected client and server events: player connection, map init, game objects init
