+++ src/TinyCiv.Client/App.xaml
