+++ src/TinyCiv.Client/App.xaml.cs
