+++ src/TinyCiv.Client/AssemblyInfo.cs
