+++ src/TinyCiv.Client/Assets/game_tile.png
