+++ src/TinyCiv.Client/Assets/warrior.png
