+++ src/TinyCiv.Client/Code/ClientManager.cs
