+++ src/TinyCiv.Client/Code/GameObject.cs
