+++ src/TinyCiv.Client/Code/structures/Structure.cs
