+++ src/TinyCiv.Client/Code/units/Cavalry.cs
