+++ src/TinyCiv.Client/Code/units/Warrior.cs
