+++ src/TinyCiv.Shared/Events/Client/MoveUnitEvent.cs
@@ -1,0 +1,4 @@
+﻿namespace TinyCiv.Shared.Events.Client;
+
+//  Requires implementation
+public record MoveUnitEvent(Guid unitId, int newRow, int newColumn) : ClientEvent;
