+++ src/TinyCiv.Shared/Events/Client/PlayerDisconnectEvent.cs
@@ -1,0 +1,4 @@
+﻿namespace TinyCiv.Shared.Events.Client;
+
+// Requires implementation
+public record PlayerDisconnectEvent(Guid playerId) : ClientEvent;
